Enqueue Noto Serif (#56)
* Enqueue Noto Serif

We're currently generating the custom font URL for Noto Serif through the gutenbergtheme_fonts_url() function, but we're not actually enqueuing the font anywhere. Therefore, the font doesn't actually load in.

This commit enqueues the font.

* Replacing mention of "Karla" with "Noto Serif"

We don't use the Karla font in this theme, so this must've been leftover from a copy/paste somewhere.